+++ src/main/kotlin/com/github/igr/sexyeditor/config/BackgroundConfiguration.kt
@@ -1,70 +1,5 @@
 package com.github.igr.sexyeditor.config
 
-enum class BackgroundPosition(val value: Int, val label: String) {
-    POSITION_TOP_LEFT(0, "Top-Left"),
-    POSITION_TOP_MIDDLE(1, "Top-Middle"),
-    POSITION_TOP_RIGHT(2, "Top-Right"),
-    POSITION_MIDDLE_LEFT(3, "Middle-Left"),
-    POSITION_CENTER(4, "Center"),
-    POSITION_MIDDLE_RIGHT(5, "Middle-Right"),
-    POSITION_BOTTOM_LEFT(6, "Bottom-Left"),
-    POSITION_BOTTOM_MIDDLE(7, "Bottom-Middle"),
-    POSITION_BOTTOM_RIGHT(8, "Bottom-Right");
-
-    companion object {
-        fun of(value: Int): BackgroundPosition {
-            return when (value) {
-                0 -> POSITION_TOP_LEFT
-                1 -> POSITION_TOP_MIDDLE
-                2 -> POSITION_TOP_RIGHT
-                3 -> POSITION_MIDDLE_LEFT
-                4 -> POSITION_CENTER
-                5 -> POSITION_MIDDLE_RIGHT
-                6 -> POSITION_BOTTOM_LEFT
-                7 -> POSITION_BOTTOM_MIDDLE
-                8 -> POSITION_BOTTOM_RIGHT
-                else -> throw IllegalArgumentException("Unknown value: $value")
-            }
-        }
-    }
-
-    fun isOnBottom(): Boolean {
-        return this == POSITION_BOTTOM_LEFT ||
-                this == POSITION_BOTTOM_MIDDLE ||
-                this == POSITION_BOTTOM_RIGHT
-    }
-
-    fun isHorizontallyCentered(): Boolean {
-        return this == POSITION_MIDDLE_LEFT ||
-                this == POSITION_CENTER ||
-                this == POSITION_MIDDLE_RIGHT
-    }
-
-    fun isOnTop(): Boolean {
-        return this == POSITION_TOP_LEFT ||
-                this == POSITION_TOP_MIDDLE ||
-                this == POSITION_TOP_RIGHT
-    }
-
-    fun isOnRightSide(): Boolean {
-        return this == POSITION_TOP_RIGHT ||
-                this == POSITION_MIDDLE_RIGHT ||
-                this == POSITION_BOTTOM_RIGHT
-    }
-
-    fun isVerticallyCentered(): Boolean {
-        return this == POSITION_TOP_MIDDLE ||
-                this == POSITION_CENTER ||
-                this == POSITION_BOTTOM_MIDDLE
-    }
-
-    fun isOnLeftSide(): Boolean {
-        return this == POSITION_TOP_LEFT ||
-                this == POSITION_MIDDLE_LEFT ||
-                this == POSITION_BOTTOM_LEFT
-    }
-}
-
 class BackgroundConfiguration {
 
     var name = "All editors"
@@ -101,12 +36,12 @@ class BackgroundConfiguration {
     var resizeValue = 90
 
     /**
-     * List of images.
+     * List of background images.
      */
     var fileNames = emptyArray<String>()
 
     /**
-     * Is the next image to load random one from the list.
+     * Is the next image to load random.
      */
     var random = false
 
@@ -116,7 +51,7 @@ class BackgroundConfiguration {
     var slideshow = false
 
     /**
-     * Pause in seconds between two slides.
+     * Pause in seconds between two slides during the slideshow.
      */
     var slideshowPause = 3
 
@@ -125,7 +60,10 @@ class BackgroundConfiguration {
      */
     var fixedPosition = true
 
-    var shrinkToFit = false
+    /**
+     * How to fit the background image.
+     */
+    var fitType = FitType.NONE
 
     override fun toString() = name
 }